10/4/2023 3:37PM

+++ Assets/Scripts/Enemy/FlyEnemySkill.cs
@@ -47,7 +47,7 @@ public class FlyEnemySkill : MonoBehaviour
             return;
         }
 
-        enemyCtrl.damageReceiver.ReceiveDamage(enemyCtrl.enemyStatus.maxHP);
+        enemyCtrl.damageReceiver.ReceiveDamage((int)enemyCtrl.enemyStatus.maxHP);
         checkExplode = true;
     }
 }
